[ISSUE-247] Extract PushHandler, FetchHandler, RpcHandler from Worker; Make OpenBlocks an AbstractMessage
