+++ common/src/main/java/com/aliyun/emr/rss/common/network/protocol/RpcRequest.java
@@ -22,10 +22,10 @@ import io.netty.buffer.ByteBuf;
 
 import com.aliyun.emr.rss.common.network.buffer.ManagedBuffer;
 import com.aliyun.emr.rss.common.network.buffer.NettyManagedBuffer;
-import com.aliyun.emr.rss.common.network.server.RpcHandler;
+import com.aliyun.emr.rss.common.network.server.BaseHandler;
 
 /**
- * A generic RPC which is handled by a remote {@link RpcHandler}.
+ * A generic RPC which is handled by a remote {@link BaseHandler}.
  * This will correspond to a single
  * {@link ResponseMessage} (either success or failure).
  */
