+++ common/src/main/java/com/aliyun/emr/rss/common/network/server/TransportServer.java
@@ -47,7 +47,7 @@ public class TransportServer implements Closeable {
 
   private final TransportContext context;
   private final TransportConf conf;
-  private final RpcHandler appRpcHandler;
+  private final BaseHandler appRpcHandler;
   private final List<TransportServerBootstrap> bootstraps;
 
   private ServerBootstrap bootstrap;
@@ -60,7 +60,7 @@ public class TransportServer implements Closeable {
     TransportContext context,
     String hostToBind,
     int portToBind,
-    RpcHandler appRpcHandler,
+    BaseHandler appRpcHandler,
     List<TransportServerBootstrap> bootstraps,
     AbstractSource source) {
     this.context = context;
@@ -88,7 +88,7 @@ public class TransportServer implements Closeable {
       TransportContext context,
       String hostToBind,
       int portToBind,
-      RpcHandler appRpcHandler,
+      BaseHandler appRpcHandler,
       List<TransportServerBootstrap> bootstraps) {
     this(context, hostToBind, portToBind, appRpcHandler, bootstraps, null);
   }
@@ -150,7 +150,7 @@ public class TransportServer implements Closeable {
     bootstrap.childHandler(new ChannelInitializer<SocketChannel>() {
       @Override
       protected void initChannel(SocketChannel ch) {
-        RpcHandler rpcHandler = appRpcHandler;
+        BaseHandler rpcHandler = appRpcHandler;
         for (TransportServerBootstrap bootstrap : bootstraps) {
           rpcHandler = bootstrap.doBootstrap(ch, rpcHandler);
         }
