+++ common/src/test/java/com/aliyun/emr/rss/common/network/TransportClientFactorySuiteJ.java
@@ -28,8 +28,8 @@ import static org.junit.Assert.*;
 
 import com.aliyun.emr.rss.common.network.client.TransportClient;
 import com.aliyun.emr.rss.common.network.client.TransportClientFactory;
+import com.aliyun.emr.rss.common.network.server.BaseHandler;
 import com.aliyun.emr.rss.common.network.server.NoOpRpcHandler;
-import com.aliyun.emr.rss.common.network.server.RpcHandler;
 import com.aliyun.emr.rss.common.network.server.TransportServer;
 import com.aliyun.emr.rss.common.network.util.ConfigProvider;
 import com.aliyun.emr.rss.common.network.util.JavaUtils;
@@ -44,7 +44,7 @@ public class TransportClientFactorySuiteJ {
   @Before
   public void setUp() {
     TransportConf conf = new TransportConf("shuffle", MapConfigProvider.EMPTY);
-    RpcHandler rpcHandler = new NoOpRpcHandler();
+    BaseHandler rpcHandler = new NoOpRpcHandler();
     context = new TransportContext(conf, rpcHandler);
     server1 = context.createServer();
     server2 = context.createServer();
@@ -69,7 +69,7 @@ public class TransportClientFactorySuiteJ {
     configMap.put("rss.shuffle.io.numConnectionsPerPeer", Integer.toString(maxConnections));
     TransportConf conf = new TransportConf("shuffle", new MapConfigProvider(configMap));
 
-    RpcHandler rpcHandler = new NoOpRpcHandler();
+    BaseHandler rpcHandler = new NoOpRpcHandler();
     TransportContext context = new TransportContext(conf, rpcHandler);
     TransportClientFactory factory = context.createClientFactory();
     Set<TransportClient> clients = Collections.synchronizedSet(
